new GreenRecommendation.tsx apply dashbaord filter
